+++ src/app/progress/Faction.tsx
@@ -3,18 +3,17 @@ import {
   DestinyFactionProgression,
   DestinyInventoryComponent,
   DestinyItemComponent,
-  DestinyProgression,
   DestinyCharacterComponent
-  } from 'bungie-api-ts/destiny2';
+} from 'bungie-api-ts/destiny2';
 import classNames from 'classnames';
 import { t } from 'i18next';
 import * as React from 'react';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions.service';
-import { bungieNetPath } from '../dim-ui/bungie-image';
 import { sum } from '../util';
 import { $state } from '../ngimport-more';
 import './faction.scss';
 import { PressTip } from '../dim-ui/press-tip';
+import FactionIcon from './FactionIcon';
 
 interface FactionProps {
   factionProgress: DestinyFactionProgression;
@@ -54,29 +53,6 @@ export function Faction(props: FactionProps) {
           }
         </div>
       </div>
-    </div>
-  );
-}
-
-export function FactionIcon(props: {
-  factionProgress: DestinyProgression;
-  factionDef: DestinyFactionDefinition;
-}) {
-  const { factionProgress, factionDef } = props;
-
-  const style = {
-    strokeDashoffset: 121.622368 - (121.622368 * factionProgress.progressToNextLevel) / factionProgress.nextLevelAt
-  };
-
-  return (
-    <div className="faction-icon">
-      <svg viewBox="0 0 48 48">
-        <image xlinkHref={bungieNetPath(factionDef.displayProperties.icon)} width="48" height="48" />
-        {factionProgress.progressToNextLevel > 0 &&
-          <polygon strokeDasharray="121.622368" style={style} fillOpacity="0" stroke="#FFF" strokeWidth="3" points="24,2.5 45.5,24 24,45.5 2.5,24" strokeLinecap="butt"/>
-        }
-      </svg>
-      <div className={classNames('item-stat', 'item-faction', { 'purchase-unlocked': factionProgress.level >= 10 })}>{factionProgress.level}</div>
     </div>
   );
 }
