+++ src/app/progress/Progress.tsx
@@ -20,11 +20,11 @@ import * as _ from 'underscore';
 import { DestinyAccount } from '../accounts/destiny-account.service';
 import { isPhonePortrait, isPhonePortraitStream } from '../mediaQueries';
 import CharacterTile, { characterIsCurrent } from './CharacterTile';
-import { Faction } from './faction';
-import { Milestone } from './milestone';
+import { Faction } from './Faction';
+import { Milestone } from './Milestone';
 import './progress.scss';
 import { ProgressProfile, reloadProgress, getProgressStream } from './progress.service';
-import Quest from './quest';
+import Quest from './Quest';
 import { isWellRested } from '../inventory/store/well-rested';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions.service';
 import { BungieImage } from '../dim-ui/bungie-image';
